# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clarity - Document Understanding API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents for current user
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: documentType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document (paste text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/analyse:
    post:
      operationId: analyseDocument
      tags: [documents]
      summary: Trigger analysis of a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis started or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Usage limit exceeded or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/analysis:
    get:
      operationId: getDocumentAnalysis
      tags: [analysis]
      summary: Get analysis for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/questions:
    get:
      operationId: listDocumentQuestions
      tags: [questions]
      summary: List Q&A for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentQuestion"
    post:
      operationId: askDocumentQuestion
      tags: [questions]
      summary: Ask a question about a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskQuestionBody"
      responses:
        "201":
          description: Question answered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentQuestion"
        "400":
          description: Usage limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard summary (wow endpoint)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        documentType:
          type: ["string", "null"]
        originalFileName:
          type: ["string", "null"]
        sourceType:
          type: string
          enum: [upload, paste]
        extractedText:
          type: ["string", "null"]
        pageCountEstimate:
          type: ["integer", "null"]
        wordCount:
          type: ["integer", "null"]
        status:
          type: string
          enum: [uploaded, extracting, analysing, completed, failed]
        isDemo:
          type: boolean
        researchField:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        confidenceLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - sourceType
        - status
        - isDemo
        - createdAt
        - updatedAt
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
        documentType:
          type: string
        researchField:
          type: string
        goal:
          type: string
      required:
        - title
        - text
    DocumentAnalysis:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentType:
          type: string
        briefSummary:
          type: string
          description: Legacy alias for bottomLine.
        plainEnglishSummary:
          type: string
          description: Legacy alias for realWorldMeaning.summary.
        bottomLine:
          type: string
        primarySummary:
          $ref: "#/components/schemas/PrimarySummary"
        suggestedQuestions:
          type: array
          items:
            type: string
        whatTheyActuallyStudied:
          type: string
        trustRating:
          $ref: "#/components/schemas/TrustRating"
        takeaway:
          type: string
        practicalRelevance:
          type: string
        whatPaperActuallyShows:
          $ref: "#/components/schemas/WhatPaperActuallyShows"
        whatItDoesNotShow:
          type: array
          items:
            type: string
        keyFindings:
          type: array
          items:
            $ref: "#/components/schemas/KeyFinding"
        evidenceQuality:
          $ref: "#/components/schemas/EvidenceQuality"
        limitationsAndGotchas:
          type: array
          items:
            $ref: "#/components/schemas/RankedLimitation"
        commonMisreadings:
          type: array
          items:
            $ref: "#/components/schemas/CommonMisreading"
        realWorldMeaning:
          $ref: "#/components/schemas/RealWorldMeaning"
        practicalUse:
          $ref: "#/components/schemas/PracticalUse"
        whoThisAppliesTo:
          $ref: "#/components/schemas/WhoThisAppliesTo"
        questionsToAskBeforeTrustingIt:
          type: array
          items:
            $ref: "#/components/schemas/QuestionToAsk"
        furtherReading:
          $ref: "#/components/schemas/FurtherReading"
        methodologySnapshot:
          $ref: "#/components/schemas/MethodologySnapshot"
        keyTerms:
          type: array
          items:
            $ref: "#/components/schemas/KeyTerm"
        missingInfo:
          type: array
          items:
            $ref: "#/components/schemas/MissingInfo"
        disclaimer:
          type: string
        confidenceLevel:
          type: string
          enum: [low, medium, high]
        confidenceNotes:
          type: ["string", "null"]
        isDemo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - documentType
        - briefSummary
        - plainEnglishSummary
        - bottomLine
        - primarySummary
        - suggestedQuestions
        - whatTheyActuallyStudied
        - trustRating
        - takeaway
        - practicalRelevance
        - whatPaperActuallyShows
        - whatItDoesNotShow
        - confidenceLevel
        - keyFindings
        - evidenceQuality
        - limitationsAndGotchas
        - commonMisreadings
        - realWorldMeaning
        - practicalUse
        - whoThisAppliesTo
        - questionsToAskBeforeTrustingIt
        - furtherReading
        - methodologySnapshot
        - keyTerms
        - missingInfo
        - disclaimer
        - isDemo
        - createdAt
    TrustRating:
      type: object
      properties:
        rating:
          type: string
          enum: [Strong, Moderate, Limited, Early / exploratory, Weak evidence, Overclaimed / misleading]
        reason:
          type: string
        confidenceLevel:
          type: string
          enum: [low, medium, high]
        citationUse:
          type: string
      required:
        - rating
        - reason
        - confidenceLevel
        - citationUse
    PrimarySummary:
      type: object
      properties:
        bottomLine:
          type: string
        insightNarrative:
          type: string
        whyThisMatters:
          type: string
        trustSignal:
          type: object
          properties:
            label:
              type: string
              enum: [Early but interesting, Directionally supportive, Promising but limited, Moderate support, Stronger evidence, Contested / uncertain]
            summary:
              type: string
            confidenceLevel:
              type: string
              enum: [low, medium, high]
          required:
            - label
            - summary
            - confidenceLevel
        mainTakeaways:
          type: array
          items:
            type: string
        suggestedQuestions:
          type: array
          items:
            type: string
      required:
        - bottomLine
        - insightNarrative
        - whyThisMatters
        - trustSignal
        - mainTakeaways
        - suggestedQuestions
    WhatPaperActuallyShows:
      type: object
      properties:
        studyType:
          type: string
        population:
          type: string
        interventionOrExposure:
          type: string
        outcomesMeasured:
          type: string
        observedResult:
          type: string
        claimType:
          type: string
          enum: [causal, correlational, theoretical, speculative]
      required:
        - studyType
        - population
        - interventionOrExposure
        - outcomesMeasured
        - observedResult
        - claimType
    KeyFinding:
      type: object
      properties:
        finding:
          type: string
        sourceInPaper:
          type: string
        populationOrSample:
          type: string
        effectDirection:
          type: string
        strengthOfSupport:
          type: string
          enum: [Strong, Moderate, Limited, Theoretical / speculative]
        plainEnglishMeaning:
          type: string
        sourceText:
          type: ["string", "null"]
      required:
        - finding
        - sourceInPaper
        - populationOrSample
        - effectDirection
        - strengthOfSupport
        - plainEnglishMeaning
    EvidenceQuality:
      type: object
      properties:
        studyType:
          type: string
        sampleSize:
          type: string
        controlsOrComparisonGroups:
          type: string
        statisticalDetail:
          type: string
        effectSizes:
          type: string
        replication:
          type: string
        publicationBiasRisk:
          type: string
        fundingConflictVisibility:
          type: string
        generalisability:
          type: string
      required:
        - studyType
        - sampleSize
        - controlsOrComparisonGroups
        - statisticalDetail
        - effectSizes
        - replication
        - publicationBiasRisk
        - fundingConflictVisibility
        - generalisability
    RankedLimitation:
      type: object
      properties:
        severity:
          type: string
          enum: [Major limitation, Important limitation, Minor / transparency note]
        limitation:
          type: string
        whyItMatters:
          type: string
        practicalConsequence:
          type: string
        whatWouldStrengthenIt:
          type: string
      required:
        - severity
        - limitation
        - whyItMatters
        - practicalConsequence
        - whatWouldStrengthenIt
    CommonMisreading:
      type: object
      properties:
        misleadingClaim:
          type: string
        whatThePaperSupports:
          type: string
        whyTheDistinctionMatters:
          type: string
      required:
        - misleadingClaim
        - whatThePaperSupports
        - whyTheDistinctionMatters
    RealWorldMeaning:
      type: object
      properties:
        summary:
          type: string
        sensibleInterpretation:
          type: string
        whatToAvoid:
          type: string
        citationGuidance:
          type: string
      required:
        - summary
        - sensibleInterpretation
        - whatToAvoid
        - citationGuidance
    PracticalUse:
      type: object
      properties:
        recommendation:
          type: string
          enum: [Act on it now, Maybe useful, low-risk, Interesting but too early, Do not act on this alone]
        reasoning:
          type: string
        caution:
          type: string
      required:
        - recommendation
        - reasoning
        - caution
    WhoThisAppliesTo:
      type: object
      properties:
        likelyAppliesTo:
          type: array
          items:
            type: string
        mayNotApplyTo:
          type: array
          items:
            type: string
        uncertainty:
          type: string
      required:
        - likelyAppliesTo
        - mayNotApplyTo
        - uncertainty
    FurtherReading:
      type: object
      properties:
        note:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - note
        - suggestions
    MethodologySnapshot:
      type: object
      properties:
        design:
          type: string
        searchSource:
          type: string
        inclusionExclusionCriteria:
          type: string
        numberOfStudiesOrParticipants:
          type: string
        analysisMethod:
          type: string
        prismaMetaAnalysisStatistics:
          type: string
      required:
        - design
        - searchSource
        - inclusionExclusionCriteria
        - numberOfStudiesOrParticipants
        - analysisMethod
        - prismaMetaAnalysisStatistics
    KeyTerm:
      type: object
      properties:
        term:
          type: string
        simpleEnglish:
          type: string
      required:
        - term
        - simpleEnglish
    MissingInfo:
      type: object
      properties:
        item:
          type: string
        whyItMatters:
          type: string
      required:
        - item
        - whyItMatters
    QuestionToAsk:
      type: object
      properties:
        question:
          type: string
        whyAskThis:
          type: string
      required:
        - question
        - whyAskThis
    DocumentQuestion:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        userId:
          type: integer
        question:
          type: string
        answer:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - userId
        - question
        - answer
        - createdAt
    AskQuestionBody:
      type: object
      properties:
        question:
          type: string
      required:
        - question
...